backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
mysql-connector-python==8.2.0

# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3

# CORS
starlette==0.35.1

# Optional: OpenAI for enhanced AI responses
# openai==1.10.0
